feat/refactor: add checks and move files
